refactor(cli): split commands/root.ts into per-command modules
Break the 2,313-LOC root.ts dispatcher into a slim entry point (175 LOC)
plus dedicated files for each subcommand and a `_helpers/` directory for
shared parsing, dispatch, and workflow utilities. Behavior, public surface,
and the existing test imports are preserved via narrow re-exports from
root.ts.
