+++ packages/cli/src/commands/_helpers/remoteAgents.ts
@@ -1,0 +1,8 @@
+import { getCliAuthProvider } from '../../runtime/supabaseAuth';
+
+export async function shouldHonorRemoteAgentPriority(
+  agentName: string,
+): Promise<boolean> {
+  if (agentName.includes(':')) return false;
+  return getCliAuthProvider().isAuthenticated();
+}
