+++ packages/cli/src/commands/agents.ts
@@ -1,0 +1,62 @@
+import { defineCommand } from 'citty';
+
+import { getVisibleAgents, loadAgents } from '@agent/index';
+import { AgentCategory } from '@agent/core/AgentDataclass';
+
+import { CliExitCode } from '../runtime/exitCodes';
+import { initCliPlatform } from '../runtime/initPlatform';
+import { writeNdjsonStdout, writeTextStdout } from '../runtime/logSinks';
+
+import { contextFromArgs } from './_helpers/context';
+import { setExitCode } from './_helpers/exitCode';
+import { GLOBAL_ARGS } from './_helpers/globalArgs';
+import type { CliContext } from '../runtime/cliContext';
+
+async function listAgents(context: CliContext): Promise<number> {
+  await initCliPlatform({
+    ...context,
+    quietLogs: true,
+    skipIncludedModelAccess: true,
+  });
+  await loadAgents({ includeRemote: false });
+  const agents = [AgentCategory.Workflow, AgentCategory.ToolUse].flatMap(
+    (category) =>
+      getVisibleAgents(category).map((agent) => ({ ...agent, category })),
+  );
+
+  if (context.outputFormat === 'json') {
+    writeTextStdout(JSON.stringify(agents, null, 2));
+    return CliExitCode.Success;
+  }
+
+  if (context.outputFormat === 'ndjson') {
+    const ts = new Date().toISOString();
+    for (const agent of agents) {
+      writeNdjsonStdout({ kind: 'agent', ts, agent });
+    }
+    return CliExitCode.Success;
+  }
+
+  for (const agent of agents) {
+    writeTextStdout(
+      `${agent.category}\t${agent.name}\t${agent.description ?? ''}`,
+    );
+  }
+  return CliExitCode.Success;
+}
+
+const agentsListCommand = defineCommand({
+  meta: { name: 'list', description: 'List available agents' },
+  args: {
+    ...GLOBAL_ARGS,
+  },
+  async run(ctx) {
+    const context = await contextFromArgs(ctx.args);
+    setExitCode(await listAgents(context));
+  },
+});
+
+export const agentsCommand = defineCommand({
+  meta: { name: 'agents', description: 'Inspect TeXRA agents' },
+  subCommands: { list: agentsListCommand },
+});
