+++ packages/cli/src/commands/chat.ts
@@ -1,0 +1,23 @@
+import { defineCommand } from 'citty';
+
+import { contextFromArgs } from './_helpers/context';
+import { setExitCode } from './_helpers/exitCode';
+import { GLOBAL_ARGS, optString } from './_helpers/globalArgs';
+
+export const chatCommand = defineCommand({
+  meta: { name: 'chat', description: 'Interactive tool-use chat session' },
+  args: {
+    ...GLOBAL_ARGS,
+    agent: { type: 'string', description: 'Tool-use agent for the session' },
+    model: { type: 'string', alias: 'm', description: 'Model for the session' },
+  },
+  async run(ctx) {
+    const context = await contextFromArgs(ctx.args);
+    const { runChat } = await import('../chat/tui/runChatTui');
+    const result = await runChat(context, {
+      agentOverride: optString(ctx.args.agent),
+      modelOverride: optString(ctx.args.model),
+    });
+    setExitCode(result.exitCode);
+  },
+});
